Refactor: move __integration tests from indicators to data-broker
