+++ packages/data-broker/jest.config.ts
@@ -8,4 +8,5 @@ import baseConfig from '../../jest.config-ts.json';
 module.exports = async () => ({
   ...baseConfig,
   rootDir: '.',
+  setupFilesAfterEnv: ['../../jest.setup.js', './jest.setup.ts'],
 });
