+++ packages/data-broker/jest.setup.ts
@@ -1,0 +1,12 @@
+/**
+ * Tupaia
+ * Copyright (c) 2017 - 2021 Beyond Essential Systems Pty Ltd
+ */
+
+import { getTestDatabase, clearAllTestData } from '@tupaia/database';
+
+afterAll(async () => {
+  const database = getTestDatabase();
+  await clearAllTestData(database);
+  await database.closeConnections();
+});
