+++ packages/data-broker/src/__tests__/__integration/indicators/fixtures/analyticArithmeticFixtures.ts
