+++ packages/indicators/jest.config.ts
@@ -8,5 +8,4 @@ import baseConfig from '../../jest.config-ts.json';
 module.exports = async () => ({
   ...baseConfig,
   rootDir: '.',
-  setupFilesAfterEnv: ['../../jest.setup.js', './jest.setup.ts'],
 });
