--- packages/indicators/jest.setup.ts
@@ -1,12 +1,0 @@
-/**
- * Tupaia
- * Copyright (c) 2017 - 2020 Beyond Essential Systems Pty Ltd
- */
-
-import { getTestDatabase, clearTestData } from '@tupaia/database';
-
-afterAll(async () => {
-  const database = getTestDatabase();
-  await clearTestData(database);
-  await database.closeConnections();
-});
