+++ packages/indicators/package.json
@@ -16,15 +16,15 @@
     "build-dev": "npm run build",
     "lint": "yarn package:lint",
     "lint:fix": "yarn lint --fix",
-    "test": "yarn package:test:withdb --runInBand",
+    "test": "yarn package:test",
     "test:coverage": "yarn test --coverage"
   },
   "dependencies": {
     "@tupaia/aggregator": "1.0.0",
-    "@tupaia/data-broker": "1.0.0",
     "@tupaia/database": "1.0.0",
     "@tupaia/expression-parser": "1.0.0",
     "@tupaia/tsutils": "1.0.0",
+    "@tupaia/types": "1.0.0",
     "@tupaia/utils": "1.0.0",
     "lodash.groupby": "^4.6.0",
     "lodash.keyby": "^4.6.0"
